huchao 20220615 up
合同管理小程序
1.合同管理 签字人改为绑定登录账号，修改按钮改为保存按钮
2.合同管理 删除增加二次确认
3.上传图片过程增加 进度提示

+++ miniprogram/packageH/page/contract_manage/contract_manage.js
@@ -286,7 +286,7 @@ Page({
           _this.tableShow(e)
 
           wx.showToast({
-            title: '修改成功！',
+            title: '保存成功！',
             icon: 'none'
           })
         },
@@ -344,41 +344,52 @@ Page({
       })
       return;
     }
-      wx.cloud.callFunction({
-        name: 'sqlServer_cw',
-        data: {
-          query: "delete from contract_manage where id='" + _this.data.id + "'"
-        },
-        success: res => {
-          _this.setData({
-            contract_code:"",
-            contract_name:"",
-            contract_type:"",
-            first_party:"",
-            second_party:"",
-            creation_date:"",
-            send_out:"",
-            id:"",
+    wx.showModal({
+      title: '提示',
+      content: '是否删除此合同？',
+      success: function(res) {
+        if (res.cancel) {
+          return;
+        }else{
+          wx.cloud.callFunction({
+            name: 'sqlServer_cw',
+            data: {
+              query: "delete from contract_manage where id='" + _this.data.id + "'"
+            },
+            success: res => {
+              _this.setData({
+                contract_code:"",
+                contract_name:"",
+                contract_type:"",
+                first_party:"",
+                second_party:"",
+                creation_date:"",
+                send_out:"",
+                id:"",
+              })
+              _this.qxShow()
+              var e = ['','','','1900/1/1','2100/12/31', _this.data.company]
+              _this.tableShow(e)
+              wx.showToast({
+                title: '删除成功！',
+                icon: 'none'
+              })
+            },
+            err: res => {
+              console.log("错误!")
+            },
+            fail: res => {
+              wx.showToast({
+                title: '请求失败！',
+                icon: 'none'
+              })
+              console.log("请求失败！")
+            }
           })
-          _this.qxShow()
-          var e = ['','','','1900/1/1','2100/12/31', _this.data.company]
-          _this.tableShow(e)
-          wx.showToast({
-            title: '删除成功！',
-            icon: 'none'
-          })
-        },
-        err: res => {
-          console.log("错误!")
-        },
-        fail: res => {
-          wx.showToast({
-            title: '请求失败！',
-            icon: 'none'
-          })
-          console.log("请求失败！")
         }
-      })
+      }
+    })
+      
   },
 
   entering:function(){
